Notice refactor -- Fix tests to match new architecture structure
- modify import statements
